gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.29
alembic==1.13.1
redis[asyncio]==5.0.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
cryptography==42.0.5
python-dotenv==1.0.0
